+++ oneflow/python/test/modules/test_expand.py
@@ -99,7 +99,7 @@ class TestModule(flow.unittest.TestCase):
         of_out = flow.tmp.expand(of_input, expand_size=expand_dim)
 
         test_case.assertTrue(np.allclose(of_out.numpy(), out_np))
-    
+
     def test_expand_v3(test_case):
         input_shape = (1, 6, 5, 3)
         expand_dim = [4, 6, 5, 3]
